requirements.txt
aiogram==3.13.1
gspread-asyncio==2.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
environs==11.0.0
python-dotenv==1.0.1